+++ src/selectors.js
@@ -1,0 +1,6 @@
+export default {
+  FILTER_SECTION: `main__filter`,
+  FILTER_INPUT: `filter__input`,
+  FILTER_LABEL: `filter__label`,
+  BOARD_TASKS: `board__tasks`,
+};
